feat: do not override missing keybinds with defaults
Now that mappings are allowed to be missing, we have no technical reason of forcing the keybinds to use defaults when they are not present in the key map file. Changed the logic to add in the default mappings only if forcing the defaults. Next up, allowing clearing the key mappings in the editor with ESC or sth similar :)